# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GramDrishti API",
    "description": "Panchayat-level forecasts and advisories. Contract: Appendix A of the guides. Every response carries data_mode; 'mock' means synthetic data.",
    "version": "0.1.1"
  },
  "paths": {
...
    "/api/v1/farmers/{farmer_id}": {
      "get": {
        "tags": [
          "farmers"
        ],
        "summary": "Farmer",
        "description": "Demo farmer profile: Panchayat, language, crops with sowing dates.",
        "operationId": "farmer_api_v1_farmers__farmer_id__get",
        "parameters": [
          {
            "name": "farmer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Farmer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Farmer"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/farmers/{farmer_id}/advice": {
      "get": {
        "tags": [
          "farmers"
        ],
        "summary": "Farmer Advice",
        "description": "Approved (or edited and approved) advisories for the farmer's Panchayat on this issue date.",
        "operationId": "farmer_advice_api_v1_farmers__farmer_id__advice_get",
        "parameters": [
          {
            "name": "farmer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Farmer Id"
            }
          },
          {
            "name": "issue_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Issue Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerAdvice"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "farmers"
        ],
        "summary": "Feedback",
        "description": "Farmer tapped \"Did it rain?\".",
        "operationId": "feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/summary": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Summary",
        "description": "Model against baselines B0, B1, B2 per variable, with 95% intervals of skill.",
        "operationId": "summary_api_v1_verification_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSummary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Not computed for real data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/reliability": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Reliability",
        "description": "Reliability diagram points for one rain event.",
        "operationId": "reliability_api_v1_verification_reliability_get",
        "parameters": [
          {
            "name": "event",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RainEvent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reliability"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Not computed for real data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/coverage": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Coverage",
        "description": "Nominal against empirical interval coverage per variable.",
        "operationId": "coverage_api_v1_verification_coverage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coverage"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Not computed for real data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verification/regions": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Regions",
        "description": "Error per held-out block.",
        "operationId": "regions_api_v1_verification_regions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Regions"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Not computed for real data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impact": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Impact",
        "description": "Replay of one decision over a season: model rule against block-forecast rule.",
        "operationId": "impact_api_v1_impact_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "monsoon_2024",
              "title": "Season"
            }
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Decision",
              "default": "spray"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Impact"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Not computed for real data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data-quality": {
      "get": {
        "tags": [
          "verification"
        ],
        "summary": "Data Quality",
        "description": "Stations reporting, missing share over 30 days, stale inputs, as of an issue date.",
        "operationId": "data_quality_api_v1_data_quality_get",
        "parameters": [
          {
            "name": "issue_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQuality"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or issue date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Advisory": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "panchayat_id": {
            "type": "string",
            "title": "Panchayat Id"
          },
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "priority": {
            "$ref": "#/components/schemas/Level"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "valid_to": {
            "type": "string",
            "format": "date",
            "title": "Valid To"
          },
          "action": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "reason": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "fallback": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "thresholds_status": {
            "$ref": "#/components/schemas/ThresholdsStatus"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "audit": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "type": "array",
            "title": "Audit"
          },
          "audio": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Lang"
            },
            "type": "object",
            "title": "Audio"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "translation_status": {
            "$ref": "#/components/schemas/TranslationStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "issue_date",
          "panchayat_id",
          "block_id",
          "crop",
          "stage",
          "category",
          "priority",
          "valid_from",
          "valid_to",
          "action",
          "reason",
          "fallback",
          "confidence",
          "evidence",
          "thresholds_status",
          "status",
          "reviewed_by",
          "reviewed_at",
          "audit",
          "audio",
          "data_mode",
          "provenance",
          "translation_status"
        ],
        "title": "Advisory"
      },
...
      "Coverage": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CoverageItem"
            },
            "type": "array",
            "title": "Items"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data_mode",
          "provenance",
          "items",
          "notes"
        ],
        "title": "Coverage"
      },
...
      "DataMode": {
        "type": "string",
        "enum": [
          "mock",
          "real"
        ],
        "title": "DataMode"
      },
      "DataQuality": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "stations_total": {
            "type": "integer",
            "title": "Stations Total"
          },
          "stations_reporting_24h": {
            "type": "integer",
            "title": "Stations Reporting 24H"
          },
          "missing_share_30d": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Share 30D"
          },
          "stale_inputs": {
            "items": {
              "$ref": "#/components/schemas/StaleInput"
            },
            "type": "array",
            "title": "Stale Inputs"
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/StationStatus"
            },
            "type": "array",
            "title": "Stations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of",
          "data_mode",
          "provenance",
          "stations_total",
          "stations_reporting_24h",
          "missing_share_30d",
          "stale_inputs",
          "stations"
        ],
        "title": "DataQuality"
      },
      "Decision": {
        "type": "string",
        "enum": [
          "spray",
          "heat_alert",
          "irrigation_wait"
        ],
        "title": "Decision"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Every 4xx/5xx body: ``{\"error\": {\"code\", \"message\"}}``."
      },
...
      "Farmer": {
        "properties": {
          "farmer_id": {
            "type": "string",
            "title": "Farmer Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "panchayat_id": {
            "type": "string",
            "title": "Panchayat Id"
          },
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "language": {
            "$ref": "#/components/schemas/Lang"
          },
          "crops": {
            "items": {
              "$ref": "#/components/schemas/FarmerCrop"
            },
            "type": "array",
            "title": "Crops"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "farmer_id",
          "name",
          "panchayat_id",
          "block_id",
          "language",
          "crops",
          "data_mode"
        ],
        "title": "Farmer"
      },
      "FarmerAdvice": {
        "properties": {
          "farmer_id": {
            "type": "string",
            "title": "Farmer Id"
          },
          "panchayat_id": {
            "type": "string",
            "title": "Panchayat Id"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "language": {
            "$ref": "#/components/schemas/Lang"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Advisory"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "farmer_id",
          "panchayat_id",
          "issue_date",
          "language",
          "data_mode",
          "items"
        ],
        "title": "FarmerAdvice"
      },
      "FarmerCrop": {
        "properties": {
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "sowing_date": {
            "type": "string",
            "format": "date",
            "title": "Sowing Date"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "area_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Fraction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "crop",
          "season",
          "sowing_date",
          "expected_harvest_date",
          "area_fraction"
        ],
        "title": "FarmerCrop"
      },
      "FeedbackRequest": {
        "properties": {
          "panchayat_id": {
            "type": "string",
            "title": "Panchayat Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "reported_rain": {
            "type": "boolean",
            "title": "Reported Rain"
          },
          "intensity": {
            "$ref": "#/components/schemas/Intensity"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "panchayat_id",
          "date",
          "reported_rain",
          "intensity",
          "channel"
        ],
        "title": "FeedbackRequest"
      },
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "stored": {
            "type": "boolean",
            "title": "Stored"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "feedback": {
            "$ref": "#/components/schemas/FeedbackRequest"
          },
          "message": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "stored",
          "received_at",
          "data_mode",
          "feedback",
          "message"
        ],
        "title": "FeedbackResponse"
      },
...
      "Impact": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "model": {
            "$ref": "#/components/schemas/DecisionCounts"
          },
          "block_baseline": {
            "$ref": "#/components/schemas/DecisionCounts"
          },
          "n_decisions": {
            "type": "integer",
            "title": "N Decisions"
          },
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpactRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "season",
          "decision",
          "data_mode",
          "provenance",
          "model",
          "block_baseline",
          "n_decisions"
        ],
        "title": "Impact"
      },
...
      "Lang": {
        "type": "string",
        "enum": [
          "en",
          "hi",
          "pa"
        ],
        "title": "Lang"
      },
...
      "RainEvent": {
        "type": "string",
        "enum": [
          "rain_ge_1mm",
          "rain_ge_2_5mm",
          "rain_ge_10mm",
          "rain_ge_35mm"
        ],
        "title": "RainEvent"
      },
...
      "Regions": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data_mode",
          "provenance",
          "method",
          "items",
          "notes"
        ],
        "title": "Regions"
      },
      "Reliability": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/RainEvent"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ReliabilityPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event",
          "data_mode",
          "provenance",
          "points",
          "notes"
        ],
        "title": "Reliability"
      },
...
      "VerificationSummary": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "variables": {
            "items": {
              "$ref": "#/components/schemas/VariableSummary"
            },
            "type": "array",
            "title": "Variables"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Events"
          },
          "block_mean_error": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Var"
            },
            "type": "object",
            "title": "Block Mean Error"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data_mode",
          "provenance",
          "method",
          "period",
          "variables",
          "events",
          "block_mean_error",
          "notes"
        ],
        "title": "VerificationSummary"
      }
    }
  }
}